move form validations into utils
